+++ packages/components/src/DataTable/stories/DataTable.story.tsx
@@ -28,7 +28,6 @@ export default {
   title: 'DataTable',
 }
 
-export * from './Actions'
 export * from './Interaction'
 export * from './Percentage'
 export * from './Responsive'